+++ packages/smarthr-ui/src/components/PageCounter/PageCounter.tsx
@@ -1,44 +1,25 @@
 import { type ComponentPropsWithoutRef, memo, useMemo } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { type DecoratorsType, useDecorators } from '../../hooks/useDecorators'
 import { Cluster } from '../Layout'
-import { RangeSeparator } from '../RangeSeparator'
 import { Text } from '../Text'
 
 type Props = {
   start: number
   end: number
   total?: number
-  decorators?: DecoratorsType<DecoratorKeyTypes>
 }
 type ElementProps = Omit<ComponentPropsWithoutRef<'div'>, keyof Props>
-
-const DECORATOR_DEFAULT_TEXTS = {
-  rangeSeparator: '–',
-  rangeSeparatorVisuallyHiddenText: 'から',
-} as const
-type DecoratorKeyTypes = keyof typeof DECORATOR_DEFAULT_TEXTS
 
 const classNameGenerator = tv({ base: 'shr-text-base' })
 
 export const PageCounter = memo<Props & ElementProps>(
-  ({ start, end, total, decorators, className, ...props }) => {
+  ({ start, end, total, className, ...props }) => {
     const actualClassName = useMemo(() => classNameGenerator({ className }), [className])
-    const decorated = useDecorators<DecoratorKeyTypes>(DECORATOR_DEFAULT_TEXTS, decorators)
-    const rangeSeparatorDecorators = useMemo(
-      () => ({
-        text: () => decorated.rangeSeparator,
-        visuallyHiddenText: () => decorated.rangeSeparatorVisuallyHiddenText,
-      }),
-      [decorated],
-    )
 
     return (
       <Cluster {...props} gap={0.25} inline align="baseline" className={actualClassName}>
-        <BoldNumber>{start}</BoldNumber>
-        <RangeSeparator decorators={rangeSeparatorDecorators} />
-        <BoldNumber>{end}</BoldNumber>
+        <BoldNumber>{start}</BoldNumber>–<BoldNumber>{end}</BoldNumber>
         <Total>{total}</Total>
       </Cluster>
     )