+++ packages/smarthr-ui/CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [78.0.0](https://github.com/kufu/smarthr-ui/compare/v77.0.0...v78.0.0) (2025-10-23)
+
+
+### ⚠ BREAKING CHANGES
+
+* **Heading,PageHeading:** size propsを追加 (#5858)
+* RangeSeparatorコンポーネントを削除する (#5900)
+* **FormControl, Fieldset:** title関連の名称をlabel/legendに、dangerously*Hidden をdangerouslyHide*に修正 (#5886)
+
+### Features
+
+* AnchorButtonでtarget="_blank"を指定した場合にアイコンを表示する ([#5879](https://github.com/kufu/smarthr-ui/issues/5879)) ([da7f12c](https://github.com/kufu/smarthr-ui/commit/da7f12cd6e67cedb36f9e97c527d2eb8f8d1deb3))
+* DateFormatterをクライアントコンポーネントとして扱うように修正 ([#5898](https://github.com/kufu/smarthr-ui/issues/5898)) ([aa5f91f](https://github.com/kufu/smarthr-ui/commit/aa5f91fc556761e65ec9704ae438a1e144489dfe))
+* **Heading,PageHeading:** size propsを追加 ([#5858](https://github.com/kufu/smarthr-ui/issues/5858)) ([d7a6b40](https://github.com/kufu/smarthr-ui/commit/d7a6b407d2d5ee3dd84a529843fb1d778d935060))
+* Text, Heading, FormControl[label], Fieldset[legend] に対してIconを表示できる属性を追加 ([#5897](https://github.com/kufu/smarthr-ui/issues/5897)) ([26b2b8c](https://github.com/kufu/smarthr-ui/commit/26b2b8c52cc40919a235d3701336c7f4ba1cef88))
+
+
+### Bug Fixes
+
+* **InputFile:** InputFileをinputの挙動に寄せ、フォーム内でデータが適切に処理できるように修正 ([#5864](https://github.com/kufu/smarthr-ui/issues/5864)) ([74219b1](https://github.com/kufu/smarthr-ui/commit/74219b1f31b90820e8c436683acbd5ca25b8e327))
+* **Textarea:** スクロール可能領域内で日本語入力中にスクロール位置が先頭に戻る問題を修正 ([#5906](https://github.com/kufu/smarthr-ui/issues/5906)) ([991664d](https://github.com/kufu/smarthr-ui/commit/991664deb6a80c568083bb8cc6efdd76699ce3c1))
+
+
+* **FormControl, Fieldset:** title関連の名称をlabel/legendに、dangerously*Hidden をdangerouslyHide*に修正 ([#5886](https://github.com/kufu/smarthr-ui/issues/5886)) ([cc5f557](https://github.com/kufu/smarthr-ui/commit/cc5f557a127d5ec0bc1bd5f66573994a8fcf95ab))
+* RangeSeparatorコンポーネントを削除する ([#5900](https://github.com/kufu/smarthr-ui/issues/5900)) ([02c847b](https://github.com/kufu/smarthr-ui/commit/02c847b93bcd35ad4ca6f6629aecf50b37b8e780))
+
 ## [77.0.0](https://github.com/kufu/smarthr-ui/compare/v76.1.1...v77.0.0) (2025-10-14)